requirements.txt
python-dotenv
opencv-python
tqdm